+++ glide/samplers/__init__.py
@@ -1,5 +1,5 @@
 from glide.samplers.active import ActiveSampler
-from glide.samplers.cluster import UniformClusterSampler
+from glide.samplers.clustered import UniformClusteredSampler
 from glide.samplers.cost_optimal import CostOptimalSampler
 from glide.samplers.cost_optimal_random import CostOptimalRandomSampler
 from glide.samplers.stratified import StratifiedSampler
@@ -10,6 +10,6 @@ __all__ = [
     "CostOptimalRandomSampler",
     "CostOptimalSampler",
     "StratifiedSampler",
-    "UniformClusterSampler",
+    "UniformClusteredSampler",
     "UniformSampler",
 ]